Moved core classes to server.
