+++ .gitignore
@@ -11,6 +11,8 @@
 /Vials.Shared.UnitTests/bin
 /Vials.Core.UnitTests/obj
 /Vials.Core.UnitTests/bin
+/Vials.Server.UnitTests/obj
+/Vials.Server.UnitTests/bin
 /Vials/Shared/obj
 /Vials/Shared/bin
 /Vials.Core/bin
