+++ Vials.Core/CoreInstaller.cs
@@ -12,10 +12,10 @@ namespace Vials.Core
     {
         public static void Install(IServiceCollection serviceCollection, IConfiguration configurationRoot)
         {
-            serviceCollection.AddScoped<ISetGenerator, SetGenerator>();
-            serviceCollection.AddScoped<IPathFinder, PathFinder>();
+            //serviceCollection.AddScoped<ISetGenerator, SetGenerator>();
+            //serviceCollection.AddScoped<IPathFinder, PathFinder>();
 
-            serviceCollection.AddScoped<Shared.ICloner, Shared.Cloner>();
+            //serviceCollection.AddScoped<Shared.ICloner, Shared.Cloner>();
         }
     }
 }
