+++ Vials.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Vials.Core.UnitTests", "Via
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Vials.Shared.UnitTests", "Vials.Shared.UnitTests\Vials.Shared.UnitTests.csproj", "{5A0ED565-C67B-4DD0-9AEB-C6F75CC6F284}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Vials.Server.UnitTests", "Vials.Server.UnitTests\Vials.Server.UnitTests.csproj", "{35CB5F16-18F6-4AEA-9862-981D0C5E8B1A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -45,6 +47,10 @@ Global
 		{5A0ED565-C67B-4DD0-9AEB-C6F75CC6F284}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5A0ED565-C67B-4DD0-9AEB-C6F75CC6F284}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5A0ED565-C67B-4DD0-9AEB-C6F75CC6F284}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35CB5F16-18F6-4AEA-9862-981D0C5E8B1A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35CB5F16-18F6-4AEA-9862-981D0C5E8B1A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35CB5F16-18F6-4AEA-9862-981D0C5E8B1A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35CB5F16-18F6-4AEA-9862-981D0C5E8B1A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
