+++ Vials/Server/Controllers/VialController.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Vials.Core;
+using Vials.Server.Utilities;
 using Vials.Shared.Objects;
 
 namespace Vials.Server.Controllers
