+++ Vials/Server/Utilities/SetGenerator.cs
@@ -5,7 +5,7 @@ using System.Text;
 using Vials.Shared.Extensions;
 using Vials.Shared.Objects;
 
-namespace Vials.Core
+namespace Vials.Server.Utilities
 {
     public class SetGenerator : ISetGenerator
     {
